Add 2nd owner to phoenix entertainment scope (#91)
My package CI scripts require the oauth token for the public wally registry. By default, a token would be created from my *personal* account. This is obviously not ideal for a variety of reasons. 

This PR adds a 2nd, non-personal automation account to the owners file for stability with our CI scripts.